AiExaminee/requirements.txt
google-cloud-vision
google-cloud-storage
google-generativeai
langdetect
pyspellchecker
googletrans==4.0.0-rc1
streamlit
